+++ src/c-ray.c
@@ -14,7 +14,7 @@
 #include "datatypes/scene.h"
 #include "utils/gitsha1.h"
 #include "utils/logging.h"
-#include "utils/filehandler.h"
+#include "utils/ioutils.h"
 #include "utils/platform/terminal.h"
 #include "utils/assert.h"
 #include "datatypes/image/texture.h"
