+++ src/utils/hashtable.c
@@ -10,11 +10,11 @@
 #include "hashtable.h"
 
 #include <inttypes.h>
-#include "../utils/logging.h"
+#include "logging.h"
 #include "../datatypes/vector.h"
 #include "assert.h"
-#include "../utils/filehandler.h"
-#include "../utils/string.h"
+#include "ioutils.h"
+#include "string.h"
 
 // Fowler-Noll-Vo hash function
 uint64_t hashDataToU64(const char *data, size_t size) {
