+++ src/utils/textbuffer.c
@@ -12,7 +12,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <stdbool.h>
-#include "filehandler.h"
+#include "ioutils.h"
 #include "logging.h"
 #include "string.h"
 #include <stdio.h>
